UI/UX V1

+++ Assets/Scripts/scoreScripts/collectibleItem.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class collectibleItem : MonoBehaviour
 {
     // Score Value
-    public int itemValue = 0;
+    public double itemValue = 0;
     // This class handles the score
     private collectibleData playerScore = new collectibleData();
